requirements.txt
pandas==2.2.0
google-cloud-bigquery==3.17.0
db-dtypes==1.2.0
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0